#355: Alarm, Marketing 파일 분리
