+++ Todoary-iOS/Todoary/Network/Service/User/AlarmService.swift
@@ -1,5 +1,5 @@
 //
-//  UserService.swift
+//  AlarmService.swift
 //  Todoary
 //
 //  Created by 박지윤 on 2023/01/10.
@@ -28,17 +28,5 @@ extension AlarmService {
     
     func getUserAlarmActiveStatus(completion: @escaping (NetworkResult<Any>) -> Void){
         requestObject(AlarmRouter.getAlarms, type: AlarmActiveStautsResultModel.self, decodingMode: .model, completion: completion)
-    }
-}
-
-class MarketingService: BaseService{
-    static let shared = MarketingService()
-    private override init() {}
-}
-
-extension MarketingService {
-    
-    func modifyMarketingAgreementStatus(request: Bool, completion: @escaping (NetworkResult<Any>) -> Void){
-        requestObjectWithEmptyResponse(MarketingRouter.patchAgreement(request: request), completion: completion)
     }
 }
