+++ Todoary-iOS/Todoary/Network/Service/User/MarketingRouter.swift
@@ -1,0 +1,12 @@
+//
+//  MarketingRouter.swift
+//  Todoary
+//
+//  Created by 박지윤 on 2023/01/16.
+//
+
+import Foundation
+
+enum MarketingRouter{
+    case patchAgreement(request: Bool)
+}
